Event alarms functionality added
When adding/editing an event, the users may add multiple event alarms, setting the exact time from the start time of the event. An application has also been created that looks for alarms into the Calindori ical files and triggers alarm notifications. The users may dismiss the alarm displayed or suspend it for a -configurable- amount of time.