Offer a week view
Week view is now available; users can manage tasks and events via a
per-week dashboard. On that dashboard, a list view displays the days of
each week. The tasks and events of each day can be seen. Upon clicking to
a task/event, a new page is opened. On that page, the task/event can be
edited or deleted. Users can navigate between weeks using the action
toolbar icons.