Remove calendar import
It is almost the same as adding an external calendar, except it is
copied to a local path.

It uses a totally different code path though. If we want to have that
feature then it should probably be an option when adding an external
calendar to copy it.

For now removing it helps us moving forward with the rest of the code.