Introduce the import-calendar-data action
Up till now, users could import a calendar file to the application. In essence, this action consisted of a calendar file to be copied to the local data location of Calindori and a calendar entity to be created, based on that file. This functionality was limiting users to use different calendars for each imported set of calendar data. With this commit, users can import data -tasks and events- from a calendar file to the active calendar. If they need the data to be imported into a separate calendar, they can just create a new one before importing.

Closes !48