Add attendees support to events
Basic attendee management features have been added to calendar events:
- Show attendee emails on event cards
- Add new attendee(s)
- Delete existing attendee(s) from an event
- Event cancellation
For attendee management to work, the active calendar should be an external ical file and the calendar owner info (name and email) should not be empty. The calendar owner will be set as the event organizer. Email notifications to attendees are not natively supported by Calindori. Nevertheless, if the external calendar file is synchronized with an external provider (e.g. Nextcloud) via vdirsyncer, attendee(s) will be informed by email whenever needed, e.g., when added to an event, when the event details are changed, in case the event is canceled etc. Moreover, the contacts found on the running system can be added as event attendees, using the KPeople abstraction layer, as long as a plugin-based source does exist (e.g. kpeoplevcard).