Offer attendee role editing
After having added an attendee to an event, the users can now edit the participation status of an attendee. E.g., they may mark some attendee(s) as optional. This feature has been added in terms of a pop-up editor, triggered by the "edit" action available in the list of the event attendees.