Move actions on calendar items
On the week and day view, the actions responsible of creating new events and tasks have been added into each day or hour. E.g., on mobile, on the day view, the users can click on the handle and reveal the actions of row 14; as soon as they click on the "New event" action button, a new event starting at 14.00 will be created.